+++ CHANGELOG.md
@@ -4,6 +4,11 @@ Based on [keepachangelog.com](https://keepachangelog.com/en/1.0.0/)
 
 ## Unversioned
 
+## Added
+
+- Ran `yarn create vite` using node `v14.19.3`
+- Changed `vite.config.ts` and `package.json` for library building using [this guide](https://jivancic.com/posts/build-a-component-library.html#components)
+
 ## [0.0.1] - 2022-07-06 Initial Intent
 
 ### Added